requirements.txt
Flask==3.0.0
groq>=0.9.0
reportlab==4.0.7
python-docx==1.1.0
python-dotenv>=1.0.0
